(test): the suites' own findings, and two counters nothing had ever driven
The suites are analysed as source, so their findings are the library's findings. Eighteen helpers
took more arguments than a call should: each is now one operand block behind a variadic macro, the
MMgr shape, so a long argument list costs one pointer instead of as many registers as it has and no
call site changed. IDEMIP_CALL is in idemip_config.h beside IDEMIP_ROUND_UP.

Two helpers had to be renamed. A function-like macro fires wherever its name meets a parenthesis,
member access included, so a helper may not share a name with a namespace entry: `prefix_in` became
`feed_prefix` and `build` became `build_frag`, or `Slaac.prefix_in(w)` would have expanded too.

Also: the twenty-three drain loops whose comma-operator condition IS the body now say so, Unity's
required empty hooks say why they are empty, three reads of a 64-bit stamp into a 32-bit local are
spelled as the narrowings they are, and a nested conditional, a multi-declarator statement and a
side effect in the right operand of && are gone.

test_clock forced IDEMIP_ENABLE_TIME_DETERMINISM with a #undef of a macro the build system owns.
It now arrives from test/CMakeLists.txt as -U then -D compile OPTIONS, which CMake writes after the
definitions, so the command line carries no redefinition for the compiler to warn about.

New coverage: the five RFC 1122 sec 3.2.2 ICMPv4 error types and the two remaining RFC 4443 ICMPv6
ones each take the RFC 2011 and RFC 2466 counter of their own, and a zone-qualified policy row
answers a lookup in its zone and is passed over in any other. Line coverage 96.7 to 97.0, branch
88.0 to 88.4, functions already 100.

+++ test/unit/base/test_checksum/test_checksum.c
@@ -12,9 +12,11 @@
 
 void setUp(void)
 {
+    // Nothing to arrange: every case builds the state it needs.
 }
 void tearDown(void)
 {
+    // Nothing to release: this suite holds no allocation, only file-scope storage.
 }
 
 // The octets RFC 1071 sec 3 tabulates.
@@ -100,13 +102,13 @@ static uint16_t reference_sum(const uint8_t *p, size_t len)
     size_t i = 0u;
     while (i + 1u < len)
     {
-        sum += (uint32_t)(((uint32_t)p[i] << 8) | (uint32_t)p[i + 1u]);
+        sum += (((uint32_t)p[i] << 8) | (uint32_t)p[i + 1u]);
         sum = (sum & 0xFFFFu) + (sum >> 16);
         i += 2u;
     }
     if (i < len)
     {
-        sum += (uint32_t)((uint32_t)p[i] << 8);
+        sum += ((uint32_t)p[i] << 8);
         sum = (sum & 0xFFFFu) + (sum >> 16);
     }
     return (uint16_t)sum;
